Refactored email sending to use a function in utils folder, it takes the receiver, subject and the email content and returns a promise.
