+++ node/routes/sendmail/index.js
@@ -1,8 +1,8 @@
 import express from 'express'
-import * as mailCtrl from '../../controllers/mailsender'
+import * as mailCtrl from '../../controllers/sendMail'
 
 const router = express.Router()
 
-router.post('/problem', mailCtrl.mailFromRegProb)
+router.post('/problem', mailCtrl.sendMailToSupport)
 
 export default router
